ad Moved Course.java to entities folder. Changed package references to account for move.
