+++ src/main/java/edu/ucsb/ucsbcslas/controllers/CourseController.java
@@ -26,7 +26,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import edu.ucsb.ucsbcslas.advice.AuthControllerAdvice;
-import edu.ucsb.ucsbcslas.models.Course;
+import edu.ucsb.ucsbcslas.entities.Course;
 import edu.ucsb.ucsbcslas.repositories.CourseRepository;
 
 @RestController
