+++ src/main/java/edu/ucsb/ucsbcslas/entities/Course.java
@@ -1,4 +1,4 @@
-package edu.ucsb.ucsbcslas.models;
+package edu.ucsb.ucsbcslas.entities;
 
 import java.util.Objects;
