+++ src/main/java/edu/ucsb/ucsbcslas/repositories/CourseRepository.java
@@ -3,7 +3,7 @@ package edu.ucsb.ucsbcslas.repositories;
 import java.util.List;
 import java.util.Optional;
 
-import edu.ucsb.ucsbcslas.models.Course;
+import edu.ucsb.ucsbcslas.entities.Course;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
